Refactor tabs into different files
