+++ tuberculosis/lib/pages/calendar_tab_page.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+
+class CalendarTabPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return const Icon(Icons.do_not_disturb_alt);
+  }
+}
