+++ tuberculosis/lib/pages/pages.dart
@@ -1,0 +1,4 @@
+export 'calendar_tab_page.dart';
+export 'faq_tab_page.dart';
+export 'information_tab_page.dart';
+export 'medication_tab_page.dart';
